work on build system and preparation for alpha release
add data files stuff to cabal build, game now installs and plays using
cabal build, but you need to rename the Paths_chaos file manually to
do a cabal build, and rename it back to compile from emacs or run as a
script

add a clear database function which allows the setup to wipe an
existing database. This gives a way for a player to easily fix a
bodged database, even when the database user doesn't have permissions
to create or drop databases.

get rid of the line endings code since it doesn't work right.